# Conflicts:
#	src/main/KeysRegenerateServlet.java
#	web/view.jsp

+++ src/main/CryptoUtils.java
@@ -12,25 +12,25 @@ import java.security.spec.PKCS8EncodedKeySpec;
 import java.security.spec.X509EncodedKeySpec;
 import java.util.Arrays;
 import java.util.Base64;
+import java.util.HashMap;
 
-public class CryptoUtils
+class CryptoUtils
 {
     private static final String ALGORITHM="AES";
     private static final String TRANSFORMATION="AES";
     private static final int FILE_SIZE_LIMIT = 536870912;
 
-    public static void encrypt(File pubKey, File inputFile, File outputFile) throws Exception
+    static void encrypt(File pubKey, File inputFile, File outputFile) throws Exception
     {
-        doCrypto(Cipher.ENCRYPT_MODE, inputFile, outputFile, pubKey);
+        doCrypto(inputFile, outputFile, pubKey);
     }
 
-    public static void decrypt(File privKey, File inputFile, File outputFile) throws Exception
+    static void decrypt(File privKey, File inputFile, File outputFile) throws Exception
     {
-        doDecrypt(Cipher.DECRYPT_MODE, inputFile, outputFile, privKey);
+        doDecrypt(inputFile, outputFile, privKey);
     }
 
-    private static void doCrypto(int cipherMode, File inputFile, File outputFile, File pubKey) throws Exception, FileUploadBase.FileSizeLimitExceededException
-    {
+    private static void doCrypto(File inputFile, File outputFile, File pubKey) throws Exception {
         try {
             if (inputFile.length() > FILE_SIZE_LIMIT) {
                 throw new FileUploadBase.FileSizeLimitExceededException("File size limit exceeded", inputFile.length(), FILE_SIZE_LIMIT);
@@ -43,7 +43,7 @@ public class CryptoUtils
             String secretKeyString = Base64.getEncoder().encodeToString(secretKey.getEncoded());
 
             Cipher cipher = Cipher.getInstance(TRANSFORMATION);
-            cipher.init(cipherMode, secretKey);
+            cipher.init(Cipher.ENCRYPT_MODE, secretKey);
 
             FileInputStream inputStream = new FileInputStream(inputFile);
             byte[] inputBytes = new byte[(int)inputFile.length()];
@@ -58,48 +58,38 @@ public class CryptoUtils
             ByteArrayOutputStream output = new ByteArrayOutputStream();
 
 
-            byte[] encryptedSymKey = rsa.encrypt(secretKey.getEncoded(), getPublicKey(pubKey));
+            byte[] encryptedSymKey = RsaKeyGenerator.encrypt(secretKey.getEncoded(), getPublicKey(pubKey));
 
             // Na zaciatok dame kryptovany sym. kluc
             output.write(encryptedSymKey);
             // Na potom subor
             output.write(outputBytes);
             // Nakoniec Mac autentifinacny tag
-            output.write(getMac(outputStream, outputBytes , secretKeyString));
+            output.write(getMac(outputBytes , secretKeyString));
 
             // Zapiseme do streamu
             outputStream.write(output.toByteArray());
 
-
-            pubKey.delete();
+            if (!pubKey.getName().equals("inputKey")) {
+                pubKey.delete();
+            }
 
             inputStream.close();
             outputStream.close();
-        } catch(
-            NoSuchPaddingException
-            | NoSuchAlgorithmException
-            | InvalidKeyException
-            |BadPaddingException
-            | IllegalBlockSizeException
-            |IOException ex
-        ) {
+        } catch(Exception ex) {
             throw new Exception("Errorencrypting/decryptingfile"+ex.getMessage(),ex);
         }
     }
 
-    public static void doDecrypt(int cipherMode, File inputFile, File outputFile, File privKey) throws Exception {
+    private static void doDecrypt(File inputFile, File outputFile, File privKey) throws Exception {
         try {
-            RsaKeyGenerator rsa= new RsaKeyGenerator();
 
-            PrivateKey myPrivate = getPrivateKey(privKey);
-
-            //Todo: get first 128 bytes from inputfile
             FileInputStream inputStream = new FileInputStream(inputFile);
             byte[] inputBytes = new byte[(int)inputFile.length()];
 
             // Load file to inputbytes
             inputStream.read(inputBytes);
-
+            //Get first 128 bytes from inputfile
             byte[] cryptedKey = Arrays.copyOfRange(inputBytes, 0, 128);
 
             //erase them from inputbytes
@@ -107,10 +97,10 @@ public class CryptoUtils
             inputBytes = Arrays.copyOfRange(inputBytes, 128, inputBytes.length);
 
             // get symmetric key from crypted data
-            Key cryptedKeyFromFile =  new SecretKeySpec(rsa.decrypt(cryptedKey, getPrivateKey(privKey)), ALGORITHM);
+            Key cryptedKeyFromFile =  new SecretKeySpec(RsaKeyGenerator.decrypt(cryptedKey, getPrivateKey(privKey)), ALGORITHM);
 
             // check mac
-            if (!checkMac(inputFile, inputBytes, Base64.getEncoder().encodeToString(cryptedKeyFromFile.getEncoded()))) {
+            if (!checkMac(inputBytes, Base64.getEncoder().encodeToString(cryptedKeyFromFile.getEncoded()))) {
                 // Throw manipulated file exception
                 throw new IOException("File manipulated");
             }
@@ -120,34 +110,58 @@ public class CryptoUtils
 
             // Init cipher to decrypt file
             Cipher decryptCipher = Cipher.getInstance(TRANSFORMATION);
-            decryptCipher.init(cipherMode, cryptedKeyFromFile);
+            decryptCipher.init(Cipher.DECRYPT_MODE, cryptedKeyFromFile);
 
             byte[] outputBytes = decryptCipher.doFinal(inputBytes);
 
             FileOutputStream outputStream = new FileOutputStream(outputFile);
             outputStream.write(outputBytes);
 
-            privKey.delete();
+            if (privKey.getName().equals("inputKey")) {
+                if (!privKey.delete()) {
+                    throw new Exception("file_delete_error");
+                }
+            }
 
             inputStream.close();
             outputStream.close();
-        } catch(
-                NoSuchPaddingException
-                        | NoSuchAlgorithmException
-                        | InvalidKeyException
-                        |BadPaddingException
-                        | IllegalBlockSizeException
-                        |IOException ex
-        ) {
+        } catch(Exception ex) {
             throw new Exception("Errorencrypting/decryptingfile"+ex.getMessage(),ex);
         }
     }
 
-    
-    public static byte[] getMac(FileOutputStream encrypted, byte[] outputBytes ,String key) throws IOException, NoSuchAlgorithmException, NoSuchPaddingException, InvalidKeyException, InvalidAlgorithmParameterException, IllegalBlockSizeException, BadPaddingException {
+    String reEncrypt(HashMap<String, String> fileData, String receiverId) throws Exception {
+        File toReEncrypt = new File(fileData.get("path"));
 
-        //Append last 16 bytes of mac
-        byte[] authTag = new byte[32];
+        if (!toReEncrypt.exists()) {
+            throw new Exception("file_not_found");
+        }
+
+        FileListManager flm = new FileListManager(fileData.get("owner_id"));
+        File decrypted = new File("temp/" + flm.getAlphaNumericString() + "." + fileData.get("mime_type"));
+        File reEncrypted = new File("temp/" + flm.getAlphaNumericString() + "." + fileData.get("mime_type") + ".enc");
+
+        if (!decrypted.createNewFile() || !reEncrypted.createNewFile()) {
+            throw new Exception("creating_tempfiles_error");
+        }
+
+        // Vyberieme si private key usera od ktoreho berieme subor
+        File privKey = flm.getUserPrivKey();
+        flm.setUserId(receiverId);
+        File myPubKey = flm.getUserPubKey();
+
+        doDecrypt(toReEncrypt, decrypted, privKey);
+        doCrypto(decrypted, reEncrypted, myPubKey);
+
+        if (!decrypted.delete()) {
+            throw new Exception("temp_delete_error");
+        }
+
+        return reEncrypted.getPath();
+    }
+
+    
+    private static byte[] getMac(byte[] outputBytes, String key) throws GeneralSecurityException {
 
         Mac mac = Mac.getInstance("HmacSHA1");
         SecretKeySpec macKey = new SecretKeySpec(key.getBytes(), "HmacSHA1");
@@ -156,18 +170,16 @@ public class CryptoUtils
         return Arrays.copyOfRange(mac.doFinal(), 0, 32);
      }
 
-    public static boolean checkMac(File encrypted, byte[] inputBytes, String keyString) throws IOException, NoSuchAlgorithmException, InvalidKeyException, IllegalBlockSizeException, BadPaddingException, InvalidAlgorithmParameterException, NoSuchPaddingException {
+    private static boolean checkMac(byte[] inputBytes, String keyString)
+            throws GeneralSecurityException {
+
         int n = 32;
         int inputBytesLength = inputBytes.length;
-        byte[] authTag = new byte[n];
-        byte[] inputFileAuthTag = new byte[n];
+        byte[] authTag;
+        byte[] inputFileAuthTag;
 
         inputFileAuthTag = Arrays.copyOfRange(inputBytes,inputBytesLength - n, inputBytesLength);
-        authTag =  getMac(null, Arrays.copyOfRange(inputBytes, 0, inputBytes.length - 32 ), keyString);
-
-        inputBytes = Arrays.copyOfRange(inputBytes, 0, inputBytesLength - n);
-        // Update array length
-        inputBytesLength = inputBytes.length;
+        authTag =  getMac(Arrays.copyOfRange(inputBytes, 0, inputBytes.length - 32 ), keyString);
 
         for(int i = 0; i < n; i++) {
             if (inputFileAuthTag[i] != authTag[i]) {
@@ -178,7 +190,7 @@ public class CryptoUtils
         return true;
     }
 
-    public static PublicKey getPublicKey(File pubKeyFile) throws InvalidKeySpecException, NoSuchAlgorithmException, IOException {
+    private static PublicKey getPublicKey(File pubKeyFile) throws InvalidKeySpecException, NoSuchAlgorithmException, IOException {
 
         byte[] keyBytes = Files.readAllBytes(pubKeyFile.toPath());
 
@@ -187,7 +199,7 @@ public class CryptoUtils
         return kf.generatePublic(spec);
     }
 
-    public static PrivateKey getPrivateKey(File privKeyFile) throws NoSuchAlgorithmException, IOException, InvalidKeySpecException {
+    private static PrivateKey getPrivateKey(File privKeyFile) throws NoSuchAlgorithmException, IOException, InvalidKeySpecException {
 
         byte[] keyBytes = Files.readAllBytes(privKeyFile.toPath());
         PKCS8EncodedKeySpec spec = new PKCS8EncodedKeySpec(keyBytes);
